Connect CDs

+++ app/Services/CDBondService.php
@@ -24,6 +24,32 @@ class CDBondService
             throw new Exception('This CD/Bond account does not belong to the specified user.');
         }
 
+        DB::beginTransaction();
+
+        try {
+            $passiveSourceUser = $user->passiveSources()->create([
+                'passive_source_id' => $this->getSource()->id,
+                'plaid_account_id' => $data['plaid_account_id'],
+            ]);
+
+            $passiveSourceUser->cdBondDetails()->create();
+        } catch (Exception $e) {
+            DB::rollBack();
+
+            throw $e;
+        }
+
+        DB::commit();
+
+        return $passiveSourceUser;
+    }
+
+    public function createManual(User $user, array $data = []): PassiveSourceUser
+    {
+        if ($user->id !== auth()->id()) {
+            throw new Exception('This CD/Bond account does not belong to the specified user.');
+        }
+
         $monthlyInterest = $this->calculateMonthlyInterest($data['amount'], $data['apy']);
 
         DB::beginTransaction();
@@ -55,7 +81,7 @@ class CDBondService
             throw new Exception('This CD/Bond account does not belong to the specified user.');
         }
 
-        $monthlyInterest = $this->calculateMonthlyInterest($data['amount'], $data['apy']);
+        $monthlyInterest = $this->calculateMonthlyInterest($passiveSourceUser->plaidAccount ? $passiveSourceUser->plaidAccount->balance : $passiveSourceUser->cdBondDetails->amount, $data['apy']);
 
         try {
             $passiveSourceUser->update(['monthly_amount' => $monthlyInterest]);
@@ -77,7 +103,19 @@ class CDBondService
             throw new Exception('This CD/Bond account does not belong to the specified user.');
         }
 
-        return $passiveSourceUser->delete();
+        if ($passiveSourceUser->plaid_account_id) {
+            if (resolve(PlaidService::class)->removeAccount($passiveSourceUser->plaidAccount->token->access_token)) {
+                $token = $passiveSourceUser->plaidAccount->token;
+                $passiveSourceUser->delete();
+                $token->delete();
+            } else {
+                throw new Exception('There was a problem contacting Plaid to remove the account.');
+            }
+        } else {
+            return $passiveSourceUser->delete();
+        }
+
+        throw new Exception('There was a problem removing your CD/Bond account.');
     }
 
     private function calculateMonthlyInterest($amount, $apy): float
